modularize step 7 - move pastiethread and storagethread, clean import
